requirements.txt
# LLM Provider SDKs
anthropic
openai==1.62.0
google-auth==2.38.0
google-genai==1.2.0

# 工具和实用程序
pydantic>=2.0.0
python-dotenv>=1.0.0
json-repair>=0.2.0

# 异步支持
aiohttp>=3.8.5
aiohappyeyeballs==2.4.6
aiosignal==1.3.2

# JSON 处理
ujson>=5.8.0

# 类型提示支持
typing-extensions>=4.5.0

# 可选依赖，用于性能优化
uvloop>=0.17.0; sys_platform != 'win32'
python-dotenv
pandas
matplotlib
tabulate

# 语言处理
language_data
langcodes==3.5.0

# 时间处理
python-dateutil>=2.8.2

